+++ docs/edge/pt-BR/learn/streaming-runtime-contract.mdx
@@ -151,7 +151,7 @@ Flows conversacionais podem transmitir um turno de usuário com `stream_turn()`:
 
 ```python
 from crewai import Flow
-from crewai.experimental.conversational import ConversationConfig, ConversationState
+from crewai.flow import ConversationConfig, ConversationState
 
 
 @ConversationConfig(llm="gpt-4o-mini", defer_trace_finalization=True)
