+++ lib/crewai/src/crewai/flow/flow.py
@@ -6,7 +6,7 @@ The implementation now lives in three modules, split by concern:
   ``@router``, ``or_`` / ``and_``) and Python Flow class projection
 - ``crewai.flow.flow_definition`` -- the serializable Flow Definition contract
 - ``crewai.flow.runtime`` -- the Flow execution engine and state
-- ``crewai.experimental.conversational_mixin`` -- experimental conversational
+- ``crewai.flow.conversational_mixin`` -- conversational
   runtime extension composed onto the public ``Flow`` class
 
 Prefer importing from those modules in new code; this module preserves the
@@ -17,7 +17,7 @@ from typing import Any, TypeVar
 
 from pydantic import BaseModel
 
-from crewai.experimental.conversational_mixin import _ConversationalMixin
+from crewai.flow.conversational_mixin import _ConversationalMixin
 from crewai.flow.dsl import and_, listen, or_, router, start
 from crewai.flow.runtime import (
     _INITIAL_STATE_CLASS_MARKER,
@@ -31,7 +31,7 @@ T = TypeVar("T", bound=dict[str, Any] | BaseModel)
 
 
 class Flow(_ConversationalMixin, RuntimeFlow[T]):
-    """Public Flow class with experimental conversational extension behavior."""
+    """Public Flow class with conversational extension behavior."""
 
 
 __all__ = [
