+++ lib/crewai/tests/test_flow_definition.py
@@ -14,12 +14,22 @@ from pydantic import BaseModel, ValidationError
 import crewai.flow.dsl as flow_dsl
 import crewai.flow.flow_definition as flow_definition
 import crewai.flow.visualization.builder as visualization_builder
-from crewai.experimental import ConversationConfig, RouterConfig
 from crewai.flow.expressions import (
     FLOW_TEMPLATE_EXPRESSION_EXAMPLES,
     FLOW_TEMPLATE_EXPRESSION_RULES,
 )
-from crewai.flow import Flow, and_, human_feedback, listen, or_, persist, router, start
+from crewai.flow import (
+    ConversationConfig,
+    Flow,
+    RouterConfig,
+    and_,
+    human_feedback,
+    listen,
+    or_,
+    persist,
+    router,
+    start,
+)
 
 
 def test_flow_public_exports_are_explicit():
